ai-engine-python/requirements.txt
fastapi
uvicorn
requests
langchain
chromadb
google-generativeai
python-dotenv
